Move benchmark code into separate file
